changed DAO method getDailyHistory to desc order. MainActivity: change logic of onButtonClicked() change main_activity layout, fragment placeholder from Fragment to FrameLayout

+++ app/src/main/java/com/example/waterapp/view/progressBar/ArcProgress.java
@@ -336,7 +336,7 @@ public class ArcProgress extends View {
         bundle.putString(INSTANCE_BOTTOM_TEXT, getBottomText());
         bundle.putFloat(INSTANCE_TEXT_SIZE, getTextSize());
         bundle.putInt(INSTANCE_TEXT_COLOR, getTextColor());
-        bundle.putFloat(INSTANCE_PROGRESS, getProgress());
+        bundle.putInt(INSTANCE_PROGRESS, getProgress());
         bundle.putInt(INSTANCE_MAX, getMax());
         bundle.putInt(INSTANCE_FINISHED_STROKE_COLOR, getFinishedStrokeColor());
         bundle.putInt(INSTANCE_UNFINISHED_STROKE_COLOR, getUnfinishedStrokeColor());
